Changed: ActionView instrumentation to use prepend instead of rewrite.
